pyproject.toml
[project]
name = "afl-result-predictor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncpg>=0.31.0",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.47",
    "xgboost>=3.2.0",
]
